smart-meal-planner/requirements.txt
flask
mysql-connector-python
openai